refactor + add logging
