+++ features/automation/common/enums/TriggersTypes.ts
