+++ features/automation/controls/AdjustSlForm.tsx
@@ -1,0 +1,8 @@
+import BigNumber from "bignumber.js";
+
+export function AdjustSlForm({ id }: { id: BigNumber }) {
+    return (
+        <>
+        </>
+    )
+}
